extraction/script__kaggle/pyproject.toml
[tool.poetry]
name = "extract-project"
version = "0.1.0"
description = ""
authors = ["Gabriel Bossardi <contato@gabrielbossardi.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
kaggle = "^1.5.12"
pandas = "^1.5.2"
sqlalchemy = "^1.4.45"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
